mobile ai guide model/museum_ai/requirements.txt
fastapi
uvicorn
pandas
openai
qdrant-client
tiktoken
python-dotenv
langchain-text-splitters
pymongo
